book-backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools"

[project]
name = "book-ingestion-pipeline"
version = "0.1.0"
description = "Web content ingestion pipeline for book website"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "cohere>=4.4.3",
    "qdrant-client>=1.7.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0"
]

[tool.setuptools.packages.find]
where = ["."]